more hpyp training and decoding

+++ src/pyp/uvector.h
@@ -4,10 +4,10 @@
 #include <vector>
 
 struct uvector_hash {
-  size_t operator()(const std::vector<unsigned>& v) const {
+  size_t operator()(const std::vector<int>& v) const {
     size_t h = v.size();
     for (auto e : v)
-      h ^= e + 0x9e3779b9 + (h<<6) + (h>>2);
+      h ^= (size_t)e + 0x9e3779b9 + (h<<6) + (h>>2);
     return h;
   }
 };
